requirements.txt
# Core data science/ML
numpy>=1.25.0
pandas>=2.1.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Deep Learning
tensorflow>=2.15.0
keras>=3.0.0
h5py>=3.10.0

# PDF/Image processing
reportlab>=4.0.0
Pillow>=10.1.0

# Web/API
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23

# Authentication/Security
bcrypt>=4.0.1
PyJWT>=2.8.0

# Utilities
python-dotenv>=1.0.0
email-validator>=2.0.0